requirements.txt
streamlit
reportlab
pandas
groq
python-dotenv